refactor(storage): delete legacy movers_persistence.rs + rename base to canonical name
PR #457 commit 5/N — final commit in the cascading pipeline-module
deletion series.

Deletes the legacy `crates/storage/src/movers_persistence.rs` (2489
LoC) which contained `StockMoversWriter` + `OptionMoversWriter` — both
unused since PR #449 DROP-migrated their target tables (`stock_movers`,
`option_movers`).

Renames `crates/storage/src/movers_base_persistence.rs` (1604 LoC) →
`crates/storage/src/movers_persistence.rs` so the canonical writer for
the `movers_1s` base table + 25 mat-views (powering the unified
`/api/movers` endpoint per PR #450) takes the canonical module name.

Updated `crates/storage/src/lib.rs`:
- `pub mod movers_base_persistence;` removed.
- Comment block explains the retirement + rename rationale.
- `pub mod movers_persistence;` now points at the renamed canonical
  writer (since the legacy file was deleted, no name collision).

Updated path-style references via bulk sed `movers_base_persistence` →
`movers_persistence`:
- `crates/api/src/handlers/movers_questdb.rs` (1 import)
- `crates/app/src/bhavcopy_pipeline.rs` (1 import for
  `QUESTDB_TABLE_MOVERS_1S`)
- `crates/app/src/main.rs` (1 call to
  `ensure_movers_tables_and_views`)
- `crates/app/src/observability.rs` (2 tracing target literals)
- `crates/storage/src/movers_writer.rs` (1 import for
  `QUESTDB_TABLE_MOVERS_1S`)

Updated test file:
- `crates/storage/tests/chaos_rescue_ring_overflow.rs` — removed the
  `StockMoversWriter` + `OptionMoversWriter` import and the 2
  test functions + 3 inline blocks that referenced them. Coverage for
  rescue-ring overflow is preserved by the `MoversBaseWriter` tests
  (now `MoversPersistenceWriter` after the rename) which exercise the
  same DB-1/DB-2 contract.

Build verification:
- `cargo check --workspace` — clean
- `cargo fmt --check` — clean
- `cargo clippy --workspace -- -D warnings -W clippy::perf` — clean
- `cargo test -p tickvault-storage --lib` — 1,564 passed
- `cargo test -p tickvault-storage --tests` — all green except 1
  pre-existing failure in `test_every_wave_5_item_carries_guarantee_matrix`
  (`active-plan-non-aws-readiness-fixes.md` is missing the 15+7 row
  guarantee matrix). Reproduced on `git stash` of my changes — fully
  unrelated to this PR. Should be fixed in a follow-up that adds
  the matrix to that plan file.

Net diff (whole PR #457 across 5 commits):
  Files deleted: 4
    crates/core/src/pipeline/top_movers.rs (-3988 LoC)
    crates/core/src/pipeline/option_movers.rs (-1031 LoC)
    crates/core/src/pipeline/movers_window.rs (-375 LoC)
    crates/storage/src/movers_persistence.rs (legacy, -2489 LoC)
  Files renamed: 1
    crates/storage/src/movers_base_persistence.rs →
    crates/storage/src/movers_persistence.rs
  Total LoC removed: ~7,883 (legacy in-memory tracker + legacy ILP
  writers — all dead code post PR #449 + PR #450).

PR #457 series complete:
- 1/N (commit 437ef39): MoverEntry relocation + dead persist call site
  + unused vars/const
- 2/N (commit 357fa93): strip 6 movers params from run_tick_processor
  + main.rs call sites + test helpers (~516 LoC removed from
  tick_processor.rs)
- 3/N (commit af1aaa6): drop SharedTopMoversSnapshot from API state +
  plumbing (105 LoC removed)
- 3.b/N (commit 1ffdcd6): fmt + clippy doc-lazy-continuation lint fixes
- 4/N (commit e2569c4): delete top_movers.rs + option_movers.rs +
  movers_window.rs + clean test refs (5394 LoC removed)
- 5/N (this commit): delete legacy storage/movers_persistence.rs +
  rename movers_base_persistence.rs (~2489 LoC removed)
