+++ crates/storage/src/movers_writer.rs
@@ -24,12 +24,12 @@ use anyhow::{Context, Result};
 use questdb::ingress::{Buffer, Sender, TimestampNanos};
 use tracing::{error, info, warn};
 
-use crate::movers_base_persistence::QUESTDB_TABLE_MOVERS_1S;
+use crate::movers_persistence::QUESTDB_TABLE_MOVERS_1S;
 use tickvault_common::config::QuestDbConfig;
 use tickvault_common::sanitize::sanitize_ilp_symbol;
 
 /// One row in the `movers_1s` base table. Mirrors the 10-column
-/// schema in `movers_base_persistence.rs`.
+/// schema in `movers_persistence.rs`.
 ///
 /// Built once per second per (security_id, segment) by
 /// `movers_pipeline`. Copy + Send so the drain task can pass
